Add half-paired problem protection
Sometimes iOS goes to sleep while accessory is in
final stages of pairing and thus iOS loses pairing,
but accessory still thinks it is paired.

This patch adds a timer after pair setup to wait for
identify function to be called. If it is not called,
pairing is removed and accessory gets back to pairing
mode.